requirements.txt
streamlit==1.41.1
openai
duckduckgo-search
typing-extensions>=4.5.0
phidata==2.7.3
composio-phidata==0.6.9
composio_core
composio==0.1.1
google-search-results==2.4.2